+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Better Likes You Layout feature (new popup setting)
+- Display photos dates on fullscreen profile photos view.
 
 ### Changed
 - Local storage log of visited profiles now includes first photo URL, view date, name, age and location.